backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
google-generativeai
python-decouple
motor
pymongo[srv]
python-dotenv
pytesseract
pdf2image
Pillow